requirements.txt
Flask==2.0.1
Werkzeug==2.0.1
psycopg2-binary==2.9.1
aiohttp==3.7.4.post0
python-dotenv==0.19.1
Flask-SocketIO==5.2.0